A Simple iOS Maps Application
This map application tracks the user’s current location and allows the
user to get directions from their current location to a given address.
The application also provides several additional features such as an
estimate of the total cost of gas and pins with traffic areas and bike
routes in Richardson.